edited SongsApiModels.ts
